--- a/modules/space/shaders/keplertrails_fs.glsl
+++ b/modules/space/shaders/keplertrails_fs.glsl
@@ -51,7 +51,7 @@ Fragment getFragment() {
     vertexDistance += 1.0;
   }
 
-  float invert = pow((1.0 - vertexDistance), trailFadeExponent);  
+  float invert = pow(1.0 - vertexDistance, trailFadeExponent);  
   float fade = clamp(invert, 0.0, 1.0);
 
   // Currently even fully transparent lines can occlude other lines, thus we discard these