## This PR adds the following
- Multithreading and a bettery system for updating which segments to render for each orbit
- Multithreading for loading of data
- Support for `Camera View Direction` in addition to existing `Camera Position Normal`

## Faster loading of data
Test with `.../scene/solarsystem/sssb/mpc.asset` (1.4 million objects). Loading times changed from `3 minutes` to `12 seconds`.

## Faster and smarter `update` function
Update function is now multithreaded and can also determine when new segments needs to be calculated, thus mostly doing minimal work for each frame (unless very high delta time).

### Compared to previous performance improvements in #3739:
**Datasets: Amor, Apollo, Outer main, Main, Inner main, Jupiter trojan, Transneptunian**
(ingame time paused)
- **Mode: Trails**
  - Before: 65 fps
  - After: 300 fps

- **Mode: Points**
  - Before: 65 fps
  - After: 350 fps

- **Mode: Trails + Points**
  - Before: 34 fps
  - After: 280 fps

### It also makes it possible to have better framerate at higher delta times:
**Datasets: Amor, Apollo, Outer main, Main, Inner main, Jupiter trojan, Transneptunian**
**Mode: Trails**
- **Before**
  - 14 days/sec: 32 fps
  - 30 days/sec: 25 fps

- **After**
  - 14 days/sec: 180 fps
  - 30 days/sec: 120 fps