@alexanderbock 2025-09-22
Removing the `RenderSize` from the main_belt_asteroid.asset as well