@alexanderbock 2025-09-26
The threaded calculation also runs when the time is paused